requirements.txt
absl-py==1.4.0
aiobotocore==2.4.2
aiohttp==3.8.4
aioitertools==0.11.0
aiosignal==1.3.1
alabaster==0.7.13
annotated-types==0.5.0
anyio==3.6.2
apispec==5.2.2
apyori==1.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.2.0
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
autoscraper==1.1.14
Babel==2.11.0
backcall==0.2.0
backports.functools-lru-cache==1.6.6
bardapi==0.1.33
beautifulsoup4==4.11.1
bert-extractive-summarizer==0.10.1
bleach==5.0.1
blis==0.7.10
boto3==1.28.31
botocore==1.31.31
breadability==0.1.20
browser-cookie3==0.19.1
bs4==0.0.1
cachetools==5.3.1
catalogue==2.0.9
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
colour==0.1.5
confection==0.1.1
contourpy==1.0.6
cryptography==38.0.4
cycler==0.11.0
cymem==2.0.7
debugpy==1.6.4
decorator==4.4.2
deep-translator==1.11.4
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.7
distlib==0.3.6
Django==4.1.3
django-allauth==0.51.0
django-crispy-forms==1.14.0
django-discover-runner==1.0
django-extensions==3.2.1
django-filter==22.1
django-import-export==3.0.1
django-role-permissions==3.1.1
django-session-csrf==0.7.1
django-session-csrf-per-view==0.5.3
django-widget-tweaks==0.2.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
dnspython==2.2.1
docopt==0.6.2
docutils==0.18.1
dtreeviz==2.2.0
easy-thumbnails==2.8.3
email-validator==1.3.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.8.1
Flask==2.2.2
Flask-AppBuilder==4.2.0
Flask-Babel==2.0.0
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
Flask-SSE==1.0.0
Flask-WTF==1.0.1
flatbuffers==23.5.26
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
fuzzywuzzy==0.18.0
gast==0.4.0
gensim==4.3.1
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.3
google-cloud-speech==2.21.0
google-cloud-translate==3.12.0
google-cloud-videointelligence==2.11.3
google-pasta==0.2.0
googleapis-common-protos==1.60.0
googletrans==4.0.0rc1
graphviz==0.20.1
greenlet==2.0.1
grpcio==1.57.0
grpcio-status==1.57.0
h11==0.14.0
h2==3.2.0
h5py==3.9.0
hpack==3.0.0
hstspreload==2023.1.1
html5lib==1.1
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
huggingface-hub==0.16.4
hyperframe==5.2.0
idna==2.10
imageio==2.31.1
imageio-ffmpeg==0.4.8
imagesize==1.4.1
imbalanced-learn==0.10.1
imblearn==0.0
importlib-resources==6.0.1
indic-nlp-library==0.92
indic-transliteration==2.3.52
inflection==0.5.1
iniconfig==2.0.0
ipykernel==6.17.1
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.10
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.23.3
jupyter_client==7.4.7
jupyter_core==5.1.0
jupyterlab==3.5.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
jupyterlab_server==2.16.3
keras==2.13.1
kiwisolver==1.4.4
langcodes==3.3.0
langdetect==1.0.9
libclang==16.0.6
lxml==4.9.2
lz4==4.3.2
mahaNLP==0.9
Markdown==3.4.1
MarkupPy==1.14
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
mlxtend==0.22.0
more-itertools==9.0.0
Morfessor==2.0.6
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.9
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.1
nlp==0.4.0
nltk==3.8.1
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.5
oauthlib==3.2.2
odfpy==1.4.1
openai==0.28.0
opencv-python==4.8.0.76
openpy==0.0.2
openpyxl==3.0.10
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
pandas==1.5.2
pandocfilters==1.5.0
parser-libraries==3.7
parso==0.8.3
pathy==0.10.2
patsy==0.5.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.4
plotly==5.14.0
pluggy==1.0.0
preshed==3.0.8
prison==0.2.1
proglog==0.1.10
prometheus-client==0.15.0
prompt-toolkit==3.0.33
proto-plus==1.22.3
protobuf==4.24.0
psutil==5.9.4
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycountry==22.3.5
pycparser==2.21
pycryptodomex==3.18.0
pydantic==2.2.1
pydantic_core==2.6.1
pydub==0.25.1
Pygments==2.13.0
PyJWT==2.6.0
pymongo==4.3.3
PyMySQL==1.0.2
pyparsing==3.0.9
pyrsistent==0.19.2
PySocks==1.7.1
pystache==0.6.0
pyTelegramBotAPI==4.10.0
pytest==7.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytube==15.0.0
pytube3==9.6.4
pytz==2022.6
pywin32==305
pywinpty==2.0.9
pyxlsb==1.0.10
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
Quandl==3.7.0
redis==5.0.1
regex==2023.6.3
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
roman==4.1
rsa==4.9
s3fs==2023.1.0
s3transfer==0.6.2
sacremoses==0.0.53
safetensors==0.3.2
scikit-learn==1.2.2
scipy==1.9.3
seaborn==0.12.1
selenium==4.8.0
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sklearn==0.0.post1
smart-open==6.3.0
sniffio==1.3.0
snowballstemmer==2.2.0
sorl-thumbnail==12.9.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SpeechRecognition==3.10.0
Sphinx==6.2.1
sphinx-argparse==0.4.0
sphinx-rtd-theme==1.2.2
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
SQLAlchemy==1.4.46
SQLAlchemy-Utils==0.39.0
sqlparse==0.4.3
srsly==2.4.7
stack-data==0.6.2
static==1.1.1
statsmodels==0.13.5
summa==1.2.0
summarizer==0.0.7
sumy==0.11.0
sympy==1.12
tablib==3.2.1
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
terminado==0.17.0
thinc==8.1.12
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
torch==2.0.1
torchaudio==2.0.2+cu117
torchvision==0.15.2
tornado==6.2
tqdm==4.65.1
traitlets==5.6.0
transformers==4.32.0
trio==0.22.0
trio-websocket==0.9.2
typer==0.9.0
typing_extensions==4.5.0
tzdata==2022.7
Unidecode==1.3.7
uritemplate==4.1.1
urllib3==1.26.13
virtualenv==20.17.0
wasabi==1.1.2
wcwidth==0.2.5
webdriver-manager==4.0.1
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.15.0
wsproto==1.2.0
WTForms==3.0.1
xgboost==1.7.5
xlrd==2.0.1
XlsxWriter==3.1.7
xlwt==1.3.0
xxhash==3.3.0
yarl==1.8.2
youtube-dl==2021.12.17
youtube-transcript-api==0.6.1
